+++ decidim-bulletin_board-app/app/models/pending_message.rb
@@ -6,7 +6,7 @@ class PendingMessage < ApplicationRecord
   belongs_to :client
   belongs_to :election
 
-  enum status: [:enqueued, :rejected, :accepted].map { |status| [status, status.to_s] }.to_h
+  enum status: [:enqueued, :rejected, :accepted]
 
   def processed?
     accepted? || rejected?