+++ decidim-bulletin_board-app/spec/commands/publish_results_spec.rb
@@ -3,22 +3,22 @@
 require "rails_helper"
 require "./spec/commands/shared/log_entry_validations"
 
-RSpec.describe OpenBallotBox do
+RSpec.describe PublishResults do
   subject { described_class.call(client, message_id, signed_data) }
 
   include_context "with a signed message"
 
   let!(:election) { create(:election, status: election_status) }
-  let(:election_status) { :ready }
+  let(:election_status) { :results }
   let(:client) { Authority.first }
-  let(:message_type) { :open_ballot_box_message }
+  let(:message_type) { :publish_results_message }
   let(:message_params) { { election: election } }
 
   it "broadcasts ok" do
     expect { subject }.to broadcast(:ok, election)
   end
 
-  it "creates the log entry for the ballot box opening" do
+  it "creates the log entry for the results publication" do
     expect { subject }.to change(LogEntry, :count).by(1)
   end
 
@@ -27,10 +27,10 @@ RSpec.describe OpenBallotBox do
   end
 
   it "changes the election status" do
-    expect { subject }.to change { Election.last.status } .from("ready") .to("vote")
+    expect { subject }.to change { Election.last.status } .from("results") .to("results_published")
   end
 
-  shared_examples "opening the ballot box fails" do
+  shared_examples "publishing results fails" do
     it "doesn't create a log entry" do
       expect { subject }.not_to change(LogEntry, :count)
     end
@@ -44,12 +44,12 @@ RSpec.describe OpenBallotBox do
     end
   end
 
-  it_behaves_like "with an invalid signed data", "opening the ballot box fails"
+  it_behaves_like "with an invalid signed data", "publishing results fails"
 
-  context "when the election status is not ready" do
-    let(:election_status) { :key_ceremony }
+  context "when the election status is not results" do
+    let(:election_status) { :tally }
 
-    it_behaves_like "opening the ballot box fails"
+    it_behaves_like "publishing results fails"
 
     it "broadcasts invalid" do
       expect { subject }.to broadcast(:invalid, "The election is not in the right step")
@@ -60,9 +60,9 @@ RSpec.describe OpenBallotBox do
     let(:client) { create(:authority, private_key: private_key) }
     let(:private_key) { generate(:private_key) }
 
-    it_behaves_like "opening the ballot box fails"
+    it_behaves_like "publishing results fails"
 
-    it "broadcast invalid" do
+    it "broadcasts invalid" do
       expect { subject }.to broadcast(:invalid, "Invalid client")
     end
   end
@@ -71,9 +71,9 @@ RSpec.describe OpenBallotBox do
     let(:client) { Trustee.first }
     let(:private_key) { Test::PrivateKeys.trustees_private_keys.first }
 
-    it_behaves_like "opening the ballot box fails"
+    it_behaves_like "publishing results fails"
 
-    it "broadcast invalid" do
+    it "broadcasts invalid" do
       expect { subject }.to broadcast(:invalid, "Invalid client")
     end
   end
@@ -81,9 +81,9 @@ RSpec.describe OpenBallotBox do
   context "when the message author is not the authority" do
     let(:extra_message_params) { { authority: create(:authority) } }
 
-    it_behaves_like "opening the ballot box fails"
+    it_behaves_like "publishing results fails"
 
-    it "broadcast invalid" do
+    it "broadcasts invalid" do
       expect { subject }.to broadcast(:invalid, "Invalid message author")
     end
   end